tools/requirements.txt
anyio==3.6.2
certifi==2023.5.7
click==8.1.3
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
psycopg2==2.9.6
python-dotenv==1.0.0
sniffio==1.3.0
